+++ packages/stg-evm-v2/devtools/config/mainnet/01/feelib-v1.usdc.config.ts
@@ -64,6 +64,7 @@ const excludedChains = new Set([
     'moonriver-mainnet',
     'opbnb-mainnet',
     'shimmer-mainnet',
+    'swell-mainnet',
     'unichain-mainnet',
     'zkconsensys-mainnet',
     'zkpolygon-mainnet',