+++ packages/stg-evm-v2/devtools/config/mainnet/01/oft-token.config.ts
@@ -10,6 +10,7 @@ import { getSafeAddress } from '../../utils'
 import {
     onApe,
     onBera,
+    onBotanix,
     onCronosevm,
     onDegen,
     onFlare,
@@ -22,6 +23,7 @@ import {
     onIslander,
     onKlaytn,
     onLightlink,
+    onNibiru,
     onPeaq,
     onPlumephoenix,
     onRarible,
@@ -33,7 +35,7 @@ import {
     onXdc,
 } from '../utils'
 
-// Both USDC and USDT now (as of 2024-12-10) have their own config files, so this file is just used for WETH Hydra deployents
+// Both USDC and USDT now (as of 2024-12-10) have their own config files, so this file is just used for WETH Hydra deployments
 
 export default async (): Promise<OmniGraphHardhat<MintableNodeConfig, unknown>> => {
     // First let's create the HardhatRuntimeEnvironment objects for all networks
@@ -59,6 +61,11 @@ export default async (): Promise<OmniGraphHardhat<MintableNodeConfig, unknown>> 
         getAssetType(EndpointId.BERA_V2_MAINNET, TokenName.ETH)
     )
     const beraETH = onBera({ contractName: beraETHContractName })
+    const botanixETHContractName = getTokenDeployName(
+        TokenName.ETH,
+        getAssetType(EndpointId.BOTANIX_V2_MAINNET, TokenName.ETH)
+    )
+    const botanixETH = onBotanix({ contractName: botanixETHContractName })
     const cronosevmETHContractName = getTokenDeployName(
         TokenName.ETH,
         getAssetType(EndpointId.CRONOSEVM_V2_MAINNET, TokenName.ETH)
@@ -123,6 +130,12 @@ export default async (): Promise<OmniGraphHardhat<MintableNodeConfig, unknown>> 
     )
     const klaytnETH = onKlaytn({ contractName: klaytnETHContractName })
 
+    const nibiruETHContractName = getTokenDeployName(
+        TokenName.ETH,
+        getAssetType(EndpointId.NIBIRU_V2_MAINNET, TokenName.ETH)
+    )
+    const nibiruETH = onNibiru({ contractName: nibiruETHContractName })
+
     const plumephoenixETHContractName = getTokenDeployName(
         TokenName.ETH,
         getAssetType(EndpointId.PLUMEPHOENIX_V2_MAINNET, TokenName.ETH)
@@ -163,6 +176,7 @@ export default async (): Promise<OmniGraphHardhat<MintableNodeConfig, unknown>> 
     const getAssetAddresses = createGetAssetAddresses(getEnvironment)
     const apeAssetAddresses = await getAssetAddresses(EndpointId.APE_V2_MAINNET, [TokenName.ETH] as const)
     const beraAssetAddresses = await getAssetAddresses(EndpointId.BERA_V2_MAINNET, [TokenName.ETH] as const)
+    const botanixAssetAddresses = await getAssetAddresses(EndpointId.BOTANIX_V2_MAINNET, [TokenName.ETH] as const)
     const cronosevmAssetAddresses = await getAssetAddresses(EndpointId.CRONOSEVM_V2_MAINNET, [
         TokenName.ETH,
         TokenName.USDC,
@@ -199,6 +213,7 @@ export default async (): Promise<OmniGraphHardhat<MintableNodeConfig, unknown>> 
         TokenName.ETH,
         TokenName.USDT,
     ] as const)
+    const nibiruAssetAddresses = await getAssetAddresses(EndpointId.NIBIRU_V2_MAINNET, [TokenName.ETH] as const)
     const plumephoenixAssetAddresses = await getAssetAddresses(EndpointId.PLUMEPHOENIX_V2_MAINNET, [
         TokenName.ETH,
         TokenName.USDC,
@@ -237,6 +252,15 @@ export default async (): Promise<OmniGraphHardhat<MintableNodeConfig, unknown>> 
                     owner: getSafeAddress(EndpointId.BERA_V2_MAINNET),
                     minters: {
                         [beraAssetAddresses.ETH]: true,
+                    },
+                },
+            },
+            {
+                contract: botanixETH,
+                config: {
+                    owner: getSafeAddress(EndpointId.BOTANIX_V2_MAINNET),
+                    minters: {
+                        [botanixAssetAddresses.ETH]: true,
                     },
                 },
             },
@@ -381,6 +405,15 @@ export default async (): Promise<OmniGraphHardhat<MintableNodeConfig, unknown>> 
                     owner: getSafeAddress(EndpointId.LIGHTLINK_V2_MAINNET),
                     minters: {
                         [lightlinkAssetAddresses.USDT]: true,
+                    },
+                },
+            },
+            {
+                contract: nibiruETH,
+                config: {
+                    owner: getSafeAddress(EndpointId.NIBIRU_V2_MAINNET),
+                    minters: {
+                        [nibiruAssetAddresses.ETH]: true,
                     },
                 },
             },