+++ packages/stg-evm-v2/devtools/config/mainnet/01/usdc-token.config.ts
@@ -13,6 +13,7 @@ import {
     onAbstract,
     onApe,
     onBera,
+    onBotanix,
     onCodex,
     onCronosevm,
     onDegen,
@@ -28,6 +29,7 @@ import {
     onIslander,
     onKlaytn,
     onLightlink,
+    onNibiru,
     onPeaq,
     onPlume,
     onPlumephoenix,
@@ -56,6 +58,9 @@ assert(usdcApeAsset.address != null, `External USDC address not found for APE`)
 
 const usdcBeraAsset = getAssetNetworkConfig(EndpointId.BERA_V2_MAINNET, TokenName.USDC)
 assert(usdcBeraAsset.address != null, `External USDC address not found for BERA`)
+
+const usdcBotanixAsset = getAssetNetworkConfig(EndpointId.BOTANIX_V2_MAINNET, TokenName.USDC)
+assert(usdcBotanixAsset.address != null, `External USDC address not found for BOTANIX`)
 
 const usdcCodexAsset = getAssetNetworkConfig(EndpointId.CODEX_V2_MAINNET, TokenName.USDC)
 assert(usdcCodexAsset.address != null, `External USDC address not found for CODEX`)
@@ -86,6 +91,9 @@ assert(usdcInkAsset.address != null, `External USDC address not found for INK`)
 
 const usdcIslanderAsset = getAssetNetworkConfig(EndpointId.ISLANDER_V2_MAINNET, TokenName.USDC)
 assert(usdcIslanderAsset.address != null, `External USDC address not found for ISLANDER`)
+
+const usdcNibiruAsset = getAssetNetworkConfig(EndpointId.NIBIRU_V2_MAINNET, TokenName.USDC)
+assert(usdcNibiruAsset.address != null, `External USDC address not found for NIBIRU`)
 
 const usdcPlumeAsset = getAssetNetworkConfig(EndpointId.PLUME_V2_MAINNET, TokenName.USDC)
 assert(usdcPlumeAsset.address != null, `External USDC address not found for PLUME`)
@@ -122,6 +130,9 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
     const beraUSDCProxy = await contractFactory(
         onBera({ contractName: 'FiatTokenProxy', address: usdcBeraAsset.address })
     )
+    const botanixUSDCProxy = await contractFactory(
+        onBotanix({ contractName: 'FiatTokenProxy', address: usdcBotanixAsset.address })
+    )
     const codexUSDCProxy = await contractFactory(
         onCodex({ contractName: 'FiatTokenProxy', address: usdcCodexAsset.address })
     )
@@ -155,6 +166,9 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
     )
     const klaytnUSDCProxy = await contractFactory(onKlaytn(proxyContract))
     const lightlinkUSDCProxy = await contractFactory(onLightlink(proxyContract))
+    const nibiruUSDCProxy = await contractFactory(
+        onNibiru({ contractName: 'FiatTokenProxy', address: usdcNibiruAsset.address })
+    )
     const peaqUSDCProxy = await contractFactory(
         onPeaq({ contractName: 'FiatTokenProxy', address: usdcPeaqAsset.address })
     )
@@ -195,6 +209,9 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
 
     const beraUSDC = onBera({ ...fiatContract, address: beraUSDCProxy.contract.address })
     const beraStargateMultisig = getSafeAddress(EndpointId.BERA_V2_MAINNET)
+
+    const botanixUSDC = onBotanix({ ...fiatContract, address: botanixUSDCProxy.contract.address })
+    const botanixStargateMultisig = getSafeAddress(EndpointId.BOTANIX_V2_MAINNET)
 
     const codexUSDC = onCodex({ ...fiatContract, address: codexUSDCProxy.contract.address })
     const codexStargateMultisig = getSafeAddress(EndpointId.CODEX_V2_MAINNET)
@@ -241,6 +258,9 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
     const lightlinkUSDC = onLightlink({ ...fiatContract, address: lightlinkUSDCProxy.contract.address })
     const lightlinkStargateMultisig = getSafeAddress(EndpointId.LIGHTLINK_V2_MAINNET)
 
+    const nibiruUSDC = onNibiru({ ...fiatContract, address: nibiruUSDCProxy.contract.address })
+    const nibiruStargateMultisig = getSafeAddress(EndpointId.NIBIRU_V2_MAINNET)
+
     const peaqUSDC = onPeaq({ ...fiatContract, address: peaqUSDCProxy.contract.address })
     const peaqStargateMultisig = getSafeAddress(EndpointId.PEAQ_V2_MAINNET)
 
@@ -280,6 +300,7 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
     const abstractAssetAddresses = await getAssetAddresses(EndpointId.ABSTRACT_V2_MAINNET, usdcAssets)
     const apeAssetAddresses = await getAssetAddresses(EndpointId.APE_V2_MAINNET, usdcAssets)
     const beraAssetAddresses = await getAssetAddresses(EndpointId.BERA_V2_MAINNET, usdcAssets)
+    const botanixAssetAddresses = await getAssetAddresses(EndpointId.BOTANIX_V2_MAINNET, usdcAssets)
     const codexAssetAddresses = await getAssetAddresses(EndpointId.CODEX_V2_MAINNET, usdcAssets)
     const cronosevmAssetAddresses = await getAssetAddresses(EndpointId.CRONOSEVM_V2_MAINNET, usdcAssets)
     const degenAssetAddresses = await getAssetAddresses(EndpointId.DEGEN_V2_MAINNET, usdcAssets)
@@ -295,6 +316,7 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
     const islanderAssetAddresses = await getAssetAddresses(EndpointId.ISLANDER_V2_MAINNET, usdcAssets)
     const klaytnAssetAddresses = await getAssetAddresses(EndpointId.KLAYTN_V2_MAINNET, usdcAssets)
     const lightlinkAssetAddresses = await getAssetAddresses(EndpointId.LIGHTLINK_V2_MAINNET, usdcAssets)
+    const nibiruAssetAddresses = await getAssetAddresses(EndpointId.NIBIRU_V2_MAINNET, usdcAssets)
     const peaqAssetAddresses = await getAssetAddresses(EndpointId.PEAQ_V2_MAINNET, usdcAssets)
     const plumeAssetAddresses = await getAssetAddresses(EndpointId.PLUME_V2_MAINNET, usdcAssets)
     const plumephoenixAssetAddresses = await getAssetAddresses(EndpointId.PLUMEPHOENIX_V2_MAINNET, usdcAssets)
@@ -345,6 +367,19 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
                     blacklister: beraStargateMultisig,
                     minters: {
                         [beraAssetAddresses.USDC]: 2n ** 256n - 1n,
+                    },
+                },
+            },
+            {
+                contract: botanixUSDC,
+                config: {
+                    owner: botanixStargateMultisig,
+                    masterMinter: botanixStargateMultisig,
+                    pauser: botanixStargateMultisig,
+                    rescuer: botanixStargateMultisig,
+                    blacklister: botanixStargateMultisig,
+                    minters: {
+                        [botanixAssetAddresses.USDC]: 2n ** 256n - 1n,
                     },
                 },
             },
@@ -540,6 +575,19 @@ export default async (): Promise<OmniGraphHardhat<USDCNodeConfig, unknown>> => {
                     blacklister: lightlinkStargateMultisig,
                     minters: {
                         [lightlinkAssetAddresses.USDC]: 2n ** 256n - 1n,
+                    },
+                },
+            },
+            {
+                contract: nibiruUSDC,
+                config: {
+                    owner: nibiruStargateMultisig,
+                    masterMinter: nibiruStargateMultisig,
+                    pauser: nibiruStargateMultisig,
+                    rescuer: nibiruStargateMultisig,
+                    blacklister: nibiruStargateMultisig,
+                    minters: {
+                        [nibiruAssetAddresses.USDC]: 2n ** 256n - 1n,
                     },
                 },
             },