+++ packages/stg-evm-v2/devtools/config/mainnet/utils.ts
@@ -12,6 +12,7 @@ export const onAvax = withEid(EndpointId.AVALANCHE_V2_MAINNET)
 export const onBase = withEid(EndpointId.BASE_V2_MAINNET)
 export const onBera = withEid(EndpointId.BERA_V2_MAINNET)
 export const onBlast = withEid(EndpointId.BLAST_V2_MAINNET)
+export const onBotanix = withEid(EndpointId.BOTANIX_V2_MAINNET)
 export const onBsc = withEid(EndpointId.BSC_V2_MAINNET)
 export const onCodex = withEid(EndpointId.CODEX_V2_MAINNET)
 export const onCoredao = withEid(EndpointId.COREDAO_V2_MAINNET)
@@ -42,6 +43,7 @@ export const onMetis = withEid(EndpointId.METIS_V2_MAINNET)
 export const onMode = withEid(EndpointId.MODE_V2_MAINNET)
 export const onMoonbeam = withEid(EndpointId.MOONBEAM_V2_MAINNET)
 export const onMoonRiver = withEid(EndpointId.MOONRIVER_V2_MAINNET)
+export const onNibiru = withEid(EndpointId.NIBIRU_V2_MAINNET)
 export const onOpbnb = withEid(EndpointId.OPBNB_V2_MAINNET)
 export const onOpt = withEid(EndpointId.OPTIMISM_V2_MAINNET)
 export const onPeaq = withEid(EndpointId.PEAQ_V2_MAINNET)
@@ -58,6 +60,7 @@ export const onSoneium = withEid(EndpointId.SONEIUM_V2_MAINNET)
 export const onSonic = withEid(EndpointId.SONIC_V2_MAINNET)
 export const onStory = withEid(EndpointId.STORY_V2_MAINNET)
 export const onSuperposition = withEid(EndpointId.SUPERPOSITION_V2_MAINNET)
+export const onSwell = withEid(EndpointId.SWELL_V2_MAINNET)
 export const onTaiko = withEid(EndpointId.TAIKO_V2_MAINNET)
 export const onTelos = withEid(EndpointId.TELOS_V2_MAINNET)
 export const onUnichain = withEid(EndpointId.UNICHAIN_V2_MAINNET)
@@ -77,6 +80,7 @@ export const chainEids = {
     'base-mainnet': EndpointId.BASE_V2_MAINNET,
     'bera-mainnet': EndpointId.BERA_V2_MAINNET,
     'blast-mainnet': EndpointId.BLAST_V2_MAINNET,
+    'botanix-mainnet': EndpointId.BOTANIX_V2_MAINNET,
     'bsc-mainnet': EndpointId.BSC_V2_MAINNET,
     'codex-mainnet': EndpointId.CODEX_V2_MAINNET,
     'coredao-mainnet': EndpointId.COREDAO_V2_MAINNET,
@@ -108,6 +112,7 @@ export const chainEids = {
     'mode-mainnet': EndpointId.MODE_V2_MAINNET,
     'moonbeam-mainnet': EndpointId.MOONBEAM_V2_MAINNET,
     'moonriver-mainnet': EndpointId.MOONRIVER_V2_MAINNET,
+    'nibiru-mainnet': EndpointId.NIBIRU_V2_MAINNET,
     'opbnb-mainnet': EndpointId.OPBNB_V2_MAINNET,
     'optimism-mainnet': EndpointId.OPTIMISM_V2_MAINNET,
     'peaq-mainnet': EndpointId.PEAQ_V2_MAINNET,
@@ -124,6 +129,7 @@ export const chainEids = {
     'sonic-mainnet': EndpointId.SONIC_V2_MAINNET,
     'story-mainnet': EndpointId.STORY_V2_MAINNET,
     'superposition-mainnet': EndpointId.SUPERPOSITION_V2_MAINNET,
+    'swell-mainnet': EndpointId.SWELL_V2_MAINNET,
     'taiko-mainnet': EndpointId.TAIKO_V2_MAINNET,
     'telos-mainnet': EndpointId.TELOS_V2_MAINNET,
     'unichain-mainnet': EndpointId.UNICHAIN_V2_MAINNET,