+++ packages/stg-definitions-v2/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @stargatefinance/stg-definitions-v2
 
+## 2.0.4
+
+### Patch Changes
+
+- 391728d: bump credits
+- 18179f7: Botanix mainnet configuration
+- 332d944: swell mainnet configuration
+- 5605211: unichain bump
+- fb17003: bump versions
+- f3ae1c2: bump taxi plumephoenix
+
+## 2.0.3
+
+### Patch Changes
+
+- 2ed591e: bump versions
+- c702833: Nibiru mainnet config
+- fb569b8: xdc mainnet config
+- 8888925: manta pacific mainnet config
+- 17b46a3: zkatana shutdown
+- f4ae7dc: Glue Gas
+
 ## 2.0.2
 
 ### Patch Changes