+++ packages/stg-error-parser/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stargatefinance/stg-error-parser
 
+## 1.1.14
+
+### Patch Changes
+
+- fb17003: bump versions
+
+## 1.1.13
+
+### Patch Changes
+
+- 2ed591e: bump versions
+
 ## 1.1.12
 
 ### Patch Changes